rd-178 split layout
