+++ packages/robindoc/src/components/elements/keylink-to-content/index.tsx
@@ -2,7 +2,7 @@ import "./keylink-to-content.scss";
 import React from "react";
 import { Keylink } from "../../ui/keylink";
 
-export const KeylinkToContent = () => {
+export const KeylinkToContent: React.FC = () => {
     return (
         <Keylink className="keylink-to-content" id="main-content">
             Skip to main content
