+++ packages/robindoc/src/components/elements/keylink-to-navigation/index.tsx
@@ -2,7 +2,7 @@ import "./keylink-to-navigation.scss";
 import React from "react";
 import { Keylink } from "../../ui/keylink";
 
-export const KeylinkToNavigation = () => {
+export const KeylinkToNavigation: React.FC = () => {
     return (
         <Keylink className="keylink-to-navigation" id="navigation">
             Return to navigation
