+++ packages/robindoc/src/index.tsx
@@ -1,5 +1,7 @@
 export { initializeRobindoc } from "./core/utils/initialize-robindoc";
-export { Layout } from "./components/elements/layout";
+export { Header } from "./components/elements/header";
+export { Footer } from "./components/elements/footer";
+export { RobinProvider } from "./components/elements/robin-provider";
 export { Main } from "./components/elements/main";
 export { KeylinkToContent } from "./components/elements/keylink-to-content";
 export { KeylinkToNavigation } from "./components/elements/keylink-to-navigation";
